requirements.txt
anthropic>=0.40.0
sentence-transformers>=3.0.0
streamlit>=1.35.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytest>=8.0.0
